\name{metabarcodeR-package}
\alias{metabarcodeR-package}
\alias{metabarcodeR}
\docType{package}
\title{
\packageTitle{metabarcodeR}
}
\description{
\packageDescription{metabarcodeR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{metabarcodeR}
\packageIndices{metabarcodeR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{metabarcodeR}

Maintainer: \packageMaintainer{metabarcodeR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
